asm_int_literal, stack_region, make stuff 64-bits
when exceeding the 32 bit boundary, something becomes slow,
so investigate...

+++ back/datatype/datatype.go
@@ -2,8 +2,8 @@ package datatype
 
 type DataType interface {
 	Name() string
-	BitSize() uint32
-	ByteSize() uint32
+	BitSize() uint64
+	ByteSize() uint64
 	Equals(DataType) bool
 }
