+++ back/datatype/datatype_pointer.go
@@ -8,11 +8,11 @@ func (typ PointerType) Name() string {
 	return "*" + typ.Pointed_type.Name()
 }
 
-func (typ PointerType) BitSize() uint32 {
+func (typ PointerType) BitSize() uint64 {
 	return PTR_SIZE * 8
 }
 
-func (typ PointerType) ByteSize() uint32 {
+func (typ PointerType) ByteSize() uint64 {
 	return PTR_SIZE
 }
